Update everyones-favorite-thing-to-build-forms.md
By the time I reached this point of the tutorial, this is what my code looked like. It was slightly confusing to have this line missing ("Did I put it in the wrong place?"). Also, more `// ...` padding.

(Not sure how the code highlight is working on the site -- that might need to be fixed if I unknowingly clobbered it in this commit)