Updated suggested node version in Prequisites
Solves the issue (#3728) I had picking the right node version, 14.0.0 wasn't working for me. It would probably be better to either keep the version number updated, or just refer to node.js LTS, as 14.0.0 was an "even" version, but it's not working either. This could also lead to a page / table showing each an every node version that's compatible with redwood, instead of making newcomers guess.