Update routing-params tutorial doc
In the tutorial, once requiring the blog post route to have a parameter, the page will not display as the following link in the boilerplate is now invalid:
```
 My default route is named <code>blogPost</code>, link to me with `
        <Link to={routes.blogPost()}>BlogPost</Link>
```
It is natural for users to test a link and click it after creating it and if they do so they are greeted with the error page with the console error ```Uncaught Error: Missing parameter 'id' for route '/blog-post/{id}'.``` . 

The error may be construed by users who are not familiar with the boilerplate code as telling them they did not set up their route params correctly.
This might be a roadblock for those following the tutorial and so the aim of this text added is to clarify what is occurring if tutorial followers run into this error. Another way could be to get followers to change the boilerplate code earlier but as we are deleting all the boilerplate in the next step I don't think this is necessary. As always open to other ways to approach